+++ full-chain/sga.h
@@ -17,7 +17,7 @@ typedef struct{
 	int columnas;
 }pos;
 
-pos *sga_gpu(float *imagen, int num_endmembers, int muestras, int lineas, int bandas, int deviceSelected, float *endmember_bandas, size_t localSize);
+pos *sga_gpu(double *imagen, int num_endmembers, int muestras, int lineas, int bandas, int deviceSelected, double *endmember_bandas, size_t localSize);
 void exitOnFail(cl_int status, const char* message);
 
 #endif
